Prevent reinstallation of an already in-use group key
Track the current GTK and IGTK that is in use and when receiving a
(possibly retransmitted) Group Message 1 or WNM-Sleep Mode Response, do
not install the given key if it is already in use. This prevents an
attacker from trying to trick the client into resetting or lowering the
sequence counter associated to the group key.

[backport to 11.0: Added WPA_IGTK_MAX_LEN manually instead of picking
the dependency]

Change-Id: If19fa9446d034723c889f42694be74dffe361167
Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@cs.kuleuven.be>